* improve tests and move the helpers
